+++ src/trace.rs
@@ -2,11 +2,11 @@ use core::panic;
 
 /// First dimension is rows, second is columns. All rows have to have the same amount of columns
 pub struct Trace {
-    pub trace: Vec<Vec<u64>>,
+    pub trace: Vec<Vec<u128>>,
 }
 
 impl Trace {
-    pub fn new(trace: Vec<Vec<u64>>) -> Self {
+    pub fn new(trace: Vec<Vec<u128>>) -> Self {
         for row in trace.iter() {
             if row.len() != 3 {
                 panic!("wrong trace format");
@@ -16,7 +16,7 @@ impl Trace {
         Trace { trace }
     }
 
-    pub fn num_of_columns(&self) -> u64 {
-        self.trace[0].len() as u64
+    pub fn num_of_columns(&self) -> u128 {
+        self.trace[0].len() as u128
     }
 }
